feat(package): implement universal package adapters and UDF hooks
Implements 12 polymorphic package format adapters for Nix, Portage, APK, Slackware, XBPS, CachyOS, Snap, Flatpak, Swupd, Eopkg, Guix, and Zypper.
Adds local (adapter-level) and global (manager-level) User-Defined Function (UDF) hooks.
Resolves E0512 pointer transmutes and standardizes alloc collections.

+++ src/container/oci_runtime.rs
@@ -58,7 +58,7 @@ impl Container for SimpleContainer {
         let len = self.name.iter().position(|&b| b == 0).unwrap_or(64);
         &self.name[..len]
     }
-    fn state(&self) -> ContainerState { unsafe { core::mem::transmute(self.state.load(Ordering::SeqCst)) } }
+    fn state(&self) -> ContainerState { unsafe { core::mem::transmute(self.state.load(Ordering::SeqCst) as u32) } }
 
     fn start(&mut self) -> Result<(), ContainerError> {
         self.state.store(ContainerState::Running as usize, Ordering::SeqCst);
