+++ src/security/integrity.rs
@@ -127,7 +127,7 @@ impl SimpleFile {
 
     pub fn get_status(&self) -> IntegrityStatus {
         unsafe {
-            core::mem::transmute(self.status.load(Ordering::SeqCst))
+            core::mem::transmute(self.status.load(Ordering::SeqCst) as u32)
         }
     }
 
@@ -191,6 +191,7 @@ pub trait IntegrityMonitor {
 
 /// Integrity statistics
 #[repr(C)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub struct IntegrityStats {
     pub total_files: usize,
     pub valid_files: usize,
