+++ src/security/pki.rs
@@ -61,7 +61,7 @@ impl SimpleCertificate {
 
 impl Certificate for SimpleCertificate {
     fn id(&self) -> CertificateID { self.id }
-    fn certificate_type(&self) -> CertificateType { unsafe { core::mem::transmute(self.certificate_type.load(Ordering::SeqCst)) } }
+    fn certificate_type(&self) -> CertificateType { unsafe { core::mem::transmute(self.certificate_type.load(Ordering::SeqCst) as u32) } }
     fn subject(&self) -> &[u8] {
         let len = self.subject.iter().position(|&b| b == 0).unwrap_or(256);
         &self.subject[..len]
